move github event subscription logic
